feat: can't delete when searching, now there's a clear search button
When searching The numbers that show the total of branches in a repository change
and show how many branches were found. Also, when clicking in the checkbox to select
all, it will select only the bracnhes showing in the search result.

If the user select all branches, search and deselect *all* the branches showing in
the search result, it'll deselect only those branches showing in the search result.

To delete, the user need to click "Clear search" to see the delete button