refactor: this is a big commit that rewrite all stores
I was having too much problemas with the old stores, so I decided to
rewrite them all. Now they are more organized and easier to maintain.
Also, i think lots of bugs were fixed with this change. I can't track
all of them, but I'm sure that the app is more stable now.